requirements.txt
# Core dependencies for Google ADK Agent
google-cloud-functions-framework>=3.0.0
google-genai>=0.1.0,<1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
ruff>=0.1.0
